Remove SC2 profile summary command
The API is structured in a way that it has to be made using OAuth2 credentials in order to fetch the internal SC2 profile identifier.

Doing a full OAuth2 flow for the discord bot would need to redirect to a web-experience that can do the sign-in and then redirect back to the bot. The bot would need to host a server to listen for the redirect and hold on to the state of the request until then.

That's a lot of complexity for not much value. So I'm removing this. Leaving the season command as that could still be useful.